separate rb-tree to h and cpp
